updated, functioning
still need to run unit tests but is fully functional with like+retweet script.
Also meaning to integrate this code with gsheets, the like+retweets tracker and tie it altogether in a macro-twitter bot and add a discord bot command where community members are able to message for their position on the leaderboard/the leaderboard itself/their current social media rewards. May also need to add multiplier with total. Output could contain subtotal, multiplier and multiplied total when sent to the user.