Initial Upload of Re-engineered SM Bot
Engineered with Classes / Modals to allow for OOP to simplify the readability and efficiency of CRUD across the tracker. In addition, this script opts into a pickleDB file instead of a CSV to reduce the need for CSV data upload/download functions - decreasing the code needed for DB interaction, boosting efficiency and readability.